requirements.txt
firebase-admin==5.2.0
Flask==2.0.2
python-dotenv==0.19.2
requests==2.27.1
tinydb==4.6.1
twilio==7.6.0